requirements.txt
appdirs==1.4.3
astroid==2.4.2
beautifulsoup4==4.9.3
bs4==0.0.1
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==4.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.6
python-arango==6.0.0
pytoml==0.1.21
redis==3.5.3
requests==2.22.0
requests-toolbelt==0.9.1
retrying==1.3.3
rq==1.5.2
six==1.14.0
soupsieve==2.0.1
toml==0.10.1
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
